Delete task restricted to the owner

+++ src/tasks/tasks.service.ts
@@ -48,8 +48,8 @@ export class TasksService {
 
     }
 
-    public async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    public async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId:user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
